refactor(tui): make ViNavigationMixin a real behavior mixin
ViNavigationMixin only held ClassVar binding lists; its docstring claimed
"behavior varies by widget type" so it provided no action implementations.
In practice four classes (HelpDialog, TaskDetailDialog, StatsDialog,
ScrollableDialogBase) each reimplemented the same six action_vi_* methods
(down/up/page_down/page_up/home/end) — byte-identical across the three tabbed
dialogs — diverging only in which scroll widget to act on.

Move the six scroll actions into the mixin, delegating to an overridable
`_get_active_scroll_widget() -> VerticalScroll | None` (default None = no-op).
Each consumer now keeps only that hook:
- HelpDialog/TaskDetailDialog/StatsDialog: keep their tab-aware getter.
- ScrollableDialogBase: thin adapter over its existing _get_scroll_widget()
  (preserves the raise-on-misconfiguration behavior).
- ViSelectOverlay: unchanged — overrides the actions with cursor-based ones.

Also delete three binding sets with zero usages: VI_ALL_BINDINGS,
VI_SCROLL_BINDINGS, VI_SCROLL_ALL_BINDINGS.

Bindings and their handlers now live together; behavior is unchanged
(help_dialog's tests call the actions directly and still pass).
